requirements.txt
glog==0.3.1
plotly==4.11.0
requests==2.22.0
imageio==2.9.0
scipy==1.5.0
pandas==1.0.5
yacs==0.1.7
torch==1.5.0
matplotlib==3.1.1
torchvision==0.6.0a0+35d732a
numpy==1.17.4
scikit_image==0.16.2
folium==0.11.0
GPSPhoto==2.2.3
imagehash==4.2.0
Pillow==8.0.1
scikit_learn==0.23.2
skimage==0.0

